requirements.txt
ultralytics>=8.0.0
torch>=2.0.0
opencv-python>=4.7.0
pyyaml>=6.0
matplotlib>=3.7.0
label-studio 